State_machines is unusable for games aparently
Upping the level size brought me to ~1650 blocks on screen, this caused
framerate to drop to ~22 fps.  After digging in it turned out that
whater state_machine is doing to the class is causing a huge slowdown.
Removing it brought me right back to 60 fps and even allowed my to
render many more objects