Limit the number of workers per job
We don't want crazy crowds.  Jobs are now limited in the number of ants
that can work them.  If an ant is on its way to a job and the job fills
it will return to wandering, allowing it to find a different available
job to work on.

This will behave where the available swarm of workers will travel
towards a job until it fills.  Also rolling around the idea of the
system finding the closest X ants and assigning them the job.  I think I
like the insect swarm behavior better